Move footer functionality to Footer module
It's a bit too much code for the high level PageMods to live in,
and seems that it could be a separate and optional feature.
